Start javasrc2cpg split
