+++ joern-cli/frontends/javasrc2cpg/src/main/scala/io/joern/javasrc2cpg/util/BindingTable.scala
@@ -1,6 +1,10 @@
 package io.joern.javasrc2cpg.util
 
 import scala.collection.mutable
+import io.shiftleft.codepropertygraph.generated.nodes.NewTypeDecl
+import io.joern.x2cpg.utils.NodeBuilders.newBindingNode
+import overflowdb.BatchedUpdate.DiffGraphBuilder
+import io.shiftleft.codepropertygraph.generated.EdgeTypes
 
 case class BindingTableEntry(name: String, signature: String, implementingMethodFullName: String)
 
@@ -33,6 +37,18 @@ trait BindingTableAdapter[InputTypeDecl, AstTypeDecl, AstMethodDecl, TypeMap] {
 }
 
 object BindingTable {
+  def createBindingNodes(diffGraph: DiffGraphBuilder, typeDeclNode: NewTypeDecl, bindingTable: BindingTable): Unit = {
+    // We only sort to get stable output.
+    val sortedEntries =
+      bindingTable.getEntries.toBuffer.sortBy((entry: BindingTableEntry) => s"${entry.name}${entry.signature}")
+
+    sortedEntries.foreach { entry =>
+      val bindingNode = newBindingNode(entry.name, entry.signature, entry.implementingMethodFullName)
+
+      diffGraph.addNode(bindingNode)
+      diffGraph.addEdge(typeDeclNode, bindingNode, EdgeTypes.BINDS)
+    }
+  }
 
   def createBindingTable[InputTypeDecl, AstTypeDecl, AstMethodDecl, TypeMap](
     typeDeclFullName: String,
